Version Update for Release
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/trunk/Modules/Spatial@13647 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ APP/service/src/test/java/eu/europa/ec/fisheries/uvms/spatial/service/SpatialServiceBeanTest.java
@@ -50,8 +50,8 @@ public class SpatialServiceBeanTest extends BaseUnitilsTest {
 
         PointType point = new PointType();
         point.setCrs(3216);
-        point.setLatitude(50);
-        point.setLongitude(-21);
+        point.setLatitude((double) 50);
+        point.setLongitude((double) -21);
         closestAreaRequest.setPoint(point);
     }
     @Test
